Add ConversationsRecyclerView to MessageFragment containing 2 recyclerviews
