+++ src/main/res/layout/fragment_messages_container.xml
@@ -11,7 +11,19 @@
         android:scrollbars="vertical"
         android:layout_above="@+id/container"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
+        android:layout_toLeftOf="@+id/recyclerviewConversations"
         />
+
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/recyclerviewConversations"
+        android:layout_width="80dp"
+        android:layout_height="match_parent"
+        android:scrollbars="vertical"
+        android:layout_above="@+id/container"
+        android:layout_alignParentRight="true"
+        />
+
 
 </RelativeLayout>
